+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/response/Context.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -16,11 +16,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
+package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response;
 
-package org.eclipse.tractusx.traceability.qualitynotification.domain.contract.contract.model;
+/**
+ * Context representation for get all policies response
+ */
+public record Context(String odrl) {
+    private static final String ODRL_VALUE = "http://www.w3.org/ns/odrl/2/";
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record EdcOperator(@JsonProperty("@id") String id) {
-
+    public static Context getDefault() {
+        return new Context(ODRL_VALUE);
+    }
 }