+++ tx-backend/src/main/resources/db/migration/V14__add_tombstone_to_assets.sql
@@ -1,0 +1,4 @@
+ALTER TABLE assets_as_planned
+    ADD COLUMN "tombstone" varchar NULL;
+ALTER TABLE assets_as_built
+    ADD COLUMN "tombstone" varchar NULL;